LayoutTests/accessibility/math-mo-fences-and-separators.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
<script src="../resources/accessibility-helper.js"></script>
</head>
<body>
<div id="content">
  <math id="math">
    <mo>*</mo>
    <mo fence="true">*</mo>
    <mo separator="true">*</mo>
    <mo fence="true" separator="true">*</mo>
    <mrow>
      <mo>(</mo> <!-- fence per the operator dictionary -->
      <mi>x</mi>
      <mo>,</mo> <!-- separator per the operator dictionary -->
      <mi>y</mi>
      <mo>;</mo> <!-- separator per the operator dictionary -->
      <mi>z</mi>
      <mo>]</mo> <!-- fence per the operator dictionary -->
    </mrow>
  </math>
</div>
<p id="description"></p>
<pre id="tree"></pre>
<div id="console"></div>
<script>
    description("This verifies the expected roles and values are exposed for mo.");

    if (window.accessibilityController) {
        dumpAccessibilityTree(accessibilityController.accessibleElementById("math"), null, 0, false, false, true);
        document.getElementById("content").style.visibility = "hidden";
    }
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
